Finished up Picture I/O. Captions aren't supported in the UI, but they are in the model and I/O.
git-svn-id: http://www.wesay.org/code/WeSay/trunk@914 b25276a9-ce15-0410-9281-991c76e3d784

committer: hatton <hatton@b25276a9-ce15-0410-9281-991c76e3d784>

+++ src/WeSay.UI/PictureControl.cs
@@ -161,7 +161,10 @@ namespace WeSay.UI
 				{
 					string old = this.GetPathToImage();
 					_fileName = "Unused_" + _fileName;
-					File.Move(old, GetPathToImage());
+					if(!File.Exists(GetPathToImage()))
+					{
+						File.Move(old, GetPathToImage());
+					}
 				}
 			}
 			catch(Exception error)
